refactor: NoteManager 타입을 생성해 관련 로직 분리하기
* NoteListViewController는 NoteManager 타입을 들고 있도록 구현
* NoteDetailViewController의 delegate을 NoteListViewController가 들고 있는 NoteManager로 지정
* NoteManager의 데이터가 변경됨을 알리는 NotificationCenter 구현
* NoteListViewController는 NoteManager의 데이터가 바뀌면 notes를 업데이트하고 테이블뷰를 reload
